+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [1.30.2](https://github.com/kunchenguid/no-mistakes/compare/v1.30.1...v1.30.2) (2026-06-21)
+
+
+### Bug Fixes
+
+* **security:** load repo config from default branch, not pushed SHA (supply-chain RCE) ([#297](https://github.com/kunchenguid/no-mistakes/issues/297)) ([21ff425](https://github.com/kunchenguid/no-mistakes/commit/21ff425c7e8c2dd3dda6149cbd47d32b8c7b5da7))
+
+## [1.30.1](https://github.com/kunchenguid/no-mistakes/compare/v1.30.0...v1.30.1) (2026-06-21)
+
+
+### Bug Fixes
+
+* **agent:** kill child process group on cancel (no orphan subprocesses) ([#300](https://github.com/kunchenguid/no-mistakes/issues/300)) ([19045f6](https://github.com/kunchenguid/no-mistakes/commit/19045f6e04251f9b746467dcc6ad372cb4658f42))
+
+## [1.30.0](https://github.com/kunchenguid/no-mistakes/compare/v1.29.1...v1.30.0) (2026-06-21)
+
+
+### Features
+
+* **git:** support GitHub fork routing ([#306](https://github.com/kunchenguid/no-mistakes/issues/306)) ([7974313](https://github.com/kunchenguid/no-mistakes/commit/79743139e5930ea1f61cee5e2733b0341a2081eb))
+
 ## [1.29.1](https://github.com/kunchenguid/no-mistakes/compare/v1.29.0...v1.29.1) (2026-06-13)